python/02-use-cases/coffee_order/requirements.txt
veadk-python[extensions]==0.5.17
veadk-python==0.5.17
agentkit-sdk-python==0.5.0
google-adk==1.21.0
qrcode==8.2
tos==2.8.7
volcengine==1.0.215